Move filesystem related code to a separate package
