+++ examples/src/main/scala/com/gu/tableversions/examples/TableLoader.scala
@@ -6,8 +6,8 @@ import cats.effect.IO
 import com.gu.tableversions.core.TableVersions.{UpdateMessage, UserId}
 import com.gu.tableversions.core.{TableDefinition, TableVersions, Version}
 import com.gu.tableversions.metastore.Metastore
-import com.gu.tableversions.spark.{VersionedDataset, VersionedFileSystem}
-import com.gu.tableversions.spark.VersionedFileSystem.ConfigKeys
+import com.gu.tableversions.spark.VersionedDataset
+import com.gu.tableversions.spark.filesystem.VersionedFileSystem
 import com.typesafe.scalalogging.LazyLogging
 import org.apache.spark.sql.{Dataset, SparkSession}
