+++ examples/src/test/scala/com/gu/tableversions/examples/DatePartitionedTableLoaderSpec.scala
@@ -9,8 +9,8 @@ import com.gu.tableversions.core.Partition.PartitionColumn
 import com.gu.tableversions.core.TableVersions._
 import com.gu.tableversions.core._
 import com.gu.tableversions.metastore.Metastore
-import com.gu.tableversions.spark.VersionedFileSystem.ConfigKeys
-import com.gu.tableversions.spark.{SparkHiveMetastore, SparkHiveSuite, VersionedFileSystem}
+import com.gu.tableversions.spark.filesystem.VersionedFileSystem.ConfigKeys
+import com.gu.tableversions.spark.{SparkHiveMetastore, SparkHiveSuite}
 import org.scalatest.{FlatSpec, Matchers}
 
 /**
