+++ spark/src/test/scala/com/gu/tableversions/spark/filesystem/VersionedFileSystemSparkSpec.scala
@@ -1,8 +1,9 @@
-package com.gu.tableversions.spark
+package com.gu.tableversions.spark.filesystem
 
 import com.gu.tableversions.core.Partition.{ColumnValue, PartitionColumn}
 import com.gu.tableversions.core.{Partition, Version}
-import com.gu.tableversions.spark.VersionedFileSystem.VersionedFileSystemConfig
+import com.gu.tableversions.spark.SparkHiveSuite
+import com.gu.tableversions.spark.filesystem.VersionedFileSystem.VersionedFileSystemConfig
 import org.apache.spark.sql.SaveMode
 import org.scalatest.{FlatSpec, Matchers}
